fix mock on testing

+++ src/Applications/use_case/_test/DeleteCommentUseCase.test.js
@@ -1,4 +1,6 @@
 const DeleteCommentUseCase = require('../DeleteCommentUseCase');
+const CommentRepository = require('../../../Domains/comments/CommentRepository');
+const ThreadRepository = require('../../../Domains/threads/ThreadRepository');
 
 describe('DeleteCommentUseCase', () => {
   it('should orchestrate the delete comment action correctly', async () => {
@@ -9,15 +11,13 @@ describe('DeleteCommentUseCase', () => {
       owner: 'user-123',
     };
 
-    const mockCommentRepository = {
-      verifyCommentExists: jest.fn().mockResolvedValue(),
-      verifyCommentOwner: jest.fn().mockResolvedValue(),
-      deleteComment: jest.fn().mockResolvedValue(),
-    };
+    const mockCommentRepository = new CommentRepository();
+    const mockThreadRepository = new ThreadRepository();
 
-    const mockThreadRepository = {
-      verifyThreadExists: jest.fn().mockResolvedValue(),
-    };
+    mockThreadRepository.verifyThreadExists = jest.fn().mockResolvedValue();
+    mockCommentRepository.verifyCommentExists = jest.fn().mockResolvedValue();
+    mockCommentRepository.verifyCommentOwner = jest.fn().mockResolvedValue();
+    mockCommentRepository.deleteComment = jest.fn().mockResolvedValue();
 
     const deleteCommentUseCase = new DeleteCommentUseCase({
       commentRepository: mockCommentRepository,
@@ -28,9 +28,13 @@ describe('DeleteCommentUseCase', () => {
     await deleteCommentUseCase.execute(useCasePayload);
 
     // Assert
-    expect(mockThreadRepository.verifyThreadExists).toHaveBeenCalledWith('thread-123');
-    expect(mockCommentRepository.verifyCommentExists).toHaveBeenCalledWith('comment-123');
-    expect(mockCommentRepository.verifyCommentOwner).toHaveBeenCalledWith('comment-123', 'user-123');
-    expect(mockCommentRepository.deleteComment).toHaveBeenCalledWith('comment-123');
+    expect(mockThreadRepository.verifyThreadExists)
+      .toHaveBeenCalledWith('thread-123');
+    expect(mockCommentRepository.verifyCommentExists)
+      .toHaveBeenCalledWith('comment-123');
+    expect(mockCommentRepository.verifyCommentOwner)
+      .toHaveBeenCalledWith('comment-123', 'user-123');
+    expect(mockCommentRepository.deleteComment)
+      .toHaveBeenCalledWith('comment-123');
   });
 });
